Include stats/mechanics of Colossal appendages
Attack and health for Colossal appendages are now included in Average Stats, as well as the specified mechanics if the appendages have them.

For example, when clicking on odds for Taunt minions, the 7 mana slot now has 3 more minions but only the stats of the Colossal appendages counts (unless the Colossal also has Taunt).